refactor: ProjectPO-money类型调整

+++ crowdfunding-member-parent/crowdfunding-member-entity/src/main/java/com/vectorx/crowdfunding/entity/po/ProjectPO.java
@@ -8,7 +8,7 @@ public class ProjectPO
 
     private String projectDescription;
 
-    private Long money;
+    private Integer money;
 
     private Integer day;
 
@@ -54,11 +54,11 @@ public class ProjectPO
         this.projectDescription = projectDescription == null ? null : projectDescription.trim();
     }
 
-    public Long getMoney() {
+    public Integer getMoney() {
         return money;
     }
 
-    public void setMoney(Long money) {
+    public void setMoney(Integer money) {
         this.money = money;
     }
